selenium/requirements.txt
fake-useragent
load_dotenv
